+++ crates/web-client/src/models/library.rs
@@ -1,0 +1,32 @@
+use miden_objects::assembly::Library as NativeLibrary;
+use wasm_bindgen::prelude::*;
+
+#[wasm_bindgen]
+pub struct Library(NativeLibrary);
+
+// CONVERSIONS
+// ================================================================================================
+
+impl From<NativeLibrary> for Library {
+    fn from(native_library: NativeLibrary) -> Self {
+        Library(native_library)
+    }
+}
+
+impl From<&NativeLibrary> for Library {
+    fn from(native_library: &NativeLibrary) -> Self {
+        Library(native_library.clone())
+    }
+}
+
+impl From<Library> for NativeLibrary {
+    fn from(library: Library) -> Self {
+        library.0
+    }
+}
+
+impl From<&Library> for NativeLibrary {
+    fn from(library: &Library) -> Self {
+        library.0.clone()
+    }
+}