+++ crates/web-client/src/models/transaction_script.rs
@@ -1,7 +1,13 @@
-use miden_objects::transaction::TransactionScript as NativeTransactionScript;
+use miden_objects::{
+    Felt as NativeFelt, Word as NativeWord,
+    transaction::TransactionScript as NativeTransactionScript,
+};
 use wasm_bindgen::prelude::*;
 
 use super::rpo_digest::RpoDigest;
+use crate::models::{
+    assembler::Assembler, transaction_script_inputs::TransactionScriptInputPairArray,
+};
 
 #[derive(Clone)]
 #[wasm_bindgen]
@@ -11,6 +17,23 @@ pub struct TransactionScript(NativeTransactionScript);
 impl TransactionScript {
     pub fn root(&self) -> RpoDigest {
         self.0.root().into()
+    }
+
+    pub fn compile(
+        script_code: &str,
+        inputs: TransactionScriptInputPairArray,
+        assembler: &Assembler,
+    ) -> Result<TransactionScript, JsValue> {
+        let native_inputs: Vec<(NativeWord, Vec<NativeFelt>)> = inputs.into();
+
+        let native = NativeTransactionScript::compile(
+            script_code,
+            native_inputs, // now the compiler knows this is a Vec<…>
+            assembler.into(),
+        )
+        .map_err(|e| JsValue::from_str(&e.to_string()))?;
+
+        Ok(TransactionScript(native))
     }
 }