+++ crates/web-client/test/mocha.global.setup.mjs
@@ -61,15 +61,22 @@ before(async () => {
     async (rpcPort, remoteProverPort) => {
       const {
         Account,
+        AccountBuilder,
+        AccountComponent,
         AccountHeader,
         AccountId,
+        AccountIdAnchor,
         AccountStorageMode,
+        AccountType,
         AdviceMap,
+        Assembler,
+        AssemblerUtils,
         AuthSecretKey,
         ConsumableNoteRecord,
         Felt,
         FeltArray,
         FungibleAsset,
+        Library,
         Note,
         NoteAndArgs,
         NoteAndArgsArray,
@@ -90,12 +97,16 @@ before(async () => {
         OutputNote,
         OutputNotesArray,
         Rpo256,
+        StorageMap,
+        StorageSlot,
         TestUtils,
         TransactionFilter,
+        TransactionKernel,
         TransactionProver,
         TransactionRequest,
         TransactionResult,
         TransactionRequestBuilder,
+        TransactionScript,
         TransactionScriptInputPair,
         TransactionScriptInputPairArray,
         Word,
@@ -110,15 +121,22 @@ before(async () => {
 
       window.client = client;
       window.Account = Account;
+      window.AccountBuilder = AccountBuilder;
+      window.AccountComponent = AccountComponent;
       window.AccountHeader = AccountHeader;
       window.AccountId = AccountId;
+      window.AccountIdAnchor = AccountIdAnchor;
       window.AccountStorageMode = AccountStorageMode;
+      window.AccountType = AccountType;
       window.AdviceMap = AdviceMap;
+      window.Assembler = Assembler;
+      window.AssemblerUtils = AssemblerUtils;
       window.AuthSecretKey = AuthSecretKey;
       window.ConsumableNoteRecord = ConsumableNoteRecord;
       window.Felt = Felt;
       window.FeltArray = FeltArray;
       window.FungibleAsset = FungibleAsset;
+      window.Library = Library;
       window.Note = Note;
       window.NoteAndArgs = NoteAndArgs;
       window.NoteAndArgsArray = NoteAndArgsArray;
@@ -139,12 +157,16 @@ before(async () => {
       window.OutputNote = OutputNote;
       window.OutputNotesArray = OutputNotesArray;
       window.Rpo256 = Rpo256;
+      window.StorageMap = StorageMap;
+      window.StorageSlot = StorageSlot;
       window.TestUtils = TestUtils;
       window.TransactionFilter = TransactionFilter;
+      window.TransactionKernel = TransactionKernel;
       window.TransactionProver = TransactionProver;
       window.TransactionRequest = TransactionRequest;
       window.TransactionResult = TransactionResult;
       window.TransactionRequestBuilder = TransactionRequestBuilder;
+      window.TransactionScript = TransactionScript;
       window.TransactionScriptInputPair = TransactionScriptInputPair;
       window.TransactionScriptInputPairArray = TransactionScriptInputPairArray;
       window.WebClient = WebClient;