+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 
 * Added support for `bech32` account IDs in the CLI (#840).
+* Added support for MASM account component libraries in Web Client (#841).
 
 ### Changes